refactored interpolation matrix to externalize covariance solve
